Use healthd for offline charge mode healthd_mode_charger.cpp from lge_awifi by @fefifofum Change-Id: I0871644ad81066ba5565b70c471b9ff9f018452a
